+++ CHANGELOG.md
@@ -5,9 +5,25 @@ The format is loosely based on [Keep a Changelog](https://keepachangelog.com/en/
 
 
 ## [Unreleased]
-- Add some other notes format instead of x
+- Add support for a notes tag - x n, or n .
 - Further options for selecting indent character
 - Helpful parsing error output in separate file
+- Possibly add settings section for number of lines in the most due/ +do tag....
+- Warn if there's impossible jumping down in indent level (by more than one at a time)
+
+
+## [0.3.0] - 2020-01-05
+### Added 
+- Add 'Most due' section at top with top 5 trees based on oldest due:yyyy-mm-dd date.
+- Add '+do' section at top with lines only in same order with '+do' flag in
+- Added various helper functions to do parsing of the tree into lists
+
+### Changed
+- Clarified comments
+
+### Fixed
+- Only indent by tabs at start of line
+- Ignore whitespace lines in indenting logic
 
 ## [0.2.0] - 2020-08-12
 ### Changed